feat(dep): add split functionality
